Add files via upload
a term calculator that can handle (), *, /, +, and -. For example, if the user enters the term
(5+8)*3/8+3, the term calculator shall calculate and output the result according to the school rules of
term calculations.
These rules are: () before * and /, * and / before + and -. Several * and / are calculated from left to
right, several + and - also from left to right.
Important: The actual algorithm must be implemented itself. The use of functions like eval in
JavaScript are not permitted.